pyproject.toml
[tool.poetry]
name = "ato2024"
version = "0.1.0"
description = ""
authors = ["Jakob Lorberblatt"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
pymysql2 = "^1.3.3"
cloudevents = "^1.11.0"
#kafka = "^1.3.5"
PyMySQL = "^1.1.1"
#clickhouse = "^0.1.6"
#izihawa-commons = "^0.0.10"
fastapi-cli = "^0.0.5"
six = "^1.16.0"
kafka-python-ng = "^2.2.3"
cryptography = "^43.0.3"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
